Changing DB Schema from chat history array into chat history sub collection
