+++ components/Default/index.ts
@@ -1,0 +1,2 @@
+export { default as ChatModal, ChatHeader } from "./ChatModal"
+export { default as Sidebar } from "./Sidebar"
